comparison-agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.28.0
supabase==2.4.6
python-dotenv==1.0.1
